Fix Error, Change Titles, Change Logo.
