+++ frontend/src/index.js
@@ -3,7 +3,6 @@ import ReactDOM from 'react-dom/client';
 import './tailwind.css';
 import App from './Component/App';
 import { AuthProvider } from './context/AuthContext';
-import { toast } from 'react-toastify';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
